Refactored lobby players into separate component
