+++ web2/src/components/pages/dashboardPage.tsx
@@ -11,7 +11,6 @@ export interface DashboardPageProps {
     user : UserResponse,
     api : ApiClient,
     setUser(user: UserResponse) : void,
-    setLobbyId(lobbyId : number) : void,
     rulesUrl : string
 }
 
@@ -31,11 +30,6 @@ export default class DashboardPage extends React.Component<DashboardPageProps> {
     private rulesOnClick() : void {
         const win = window.open(this.props.rulesUrl, '_blank');
         win.focus();
-    }
-
-    componentDidMount() {
-        //Clear lobbyId from application state
-        this.props.setLobbyId(null);
     }
 
     render() {
